% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_room_outside_text.R
\name{get_room_outside_text}
\alias{get_room_outside_text}
\title{Get the text for the room called \code{outside}}
\usage{
get_room_outside_text()
}
\description{
Get the text for the room called \code{outside}
}
