% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{room_description}
\alias{room_description}
\title{Global variables}
\format{
An object of class \code{character} of length 1.
}
\usage{
room_description
}
\description{
Global variables
}
\keyword{datasets}
