% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server.R
\name{get_server}
\alias{get_server}
\title{Function that returns the server part of the app}
\usage{
get_server(input, output, session)
}
\arguments{
\item{input}{something}

\item{output}{something}

\item{session}{something}
}
\description{
Function that returns the server part of the app
}
